+++ lib/natty-ui/features.rb
@@ -10,7 +10,8 @@ module NattyUI
     # @param [#to_s] symbol string to build the horizontal rule
     # @return [Wrapper, Wrapper::Element] itself
     def hr(symbol = '═')
-      size = NattyUI.display_width(NattyUI.plain(symbol.to_s))
+      symbol = symbol.to_s
+      size = _plain_width(symbol)
       return self if size.zero?
       msg = symbol * ((available_width - 1) / size)
       return puts(msg, prefix: Ansi[39], suffix: Ansi.reset) if wrapper.ansi?
@@ -18,6 +19,9 @@ module NattyUI
     end
 
     protected
+
+    def _plain_width(str) = NattyUI.display_width(NattyUI.plain(str))
+    def _blemish_width(str) = NattyUI.display_width(Ansi.blemish(str))
 
     def _element(type, *args)
       wrapper.class.const_get(type).__send__(:new, self).__send__(:_call, *args)
