+++ lib/natty-ui/wrapper/task.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'section'
-require_relative 'mixins'
+require_relative '../mixins'
 
 module NattyUI
   module Features
